+++ src/tasks/queries/getDashboardTasks.ts
@@ -1,0 +1,45 @@
+import { resolver } from "@blitzjs/rpc"
+import { Ctx } from "blitz"
+import { TaskStatus } from "db"
+import moment from "moment"
+
+import getTasks from "./getTasks"
+
+export default resolver.pipe(resolver.authorize(), async (undefined, ctx: Ctx) => {
+  const today = moment().startOf("day")
+
+  const { tasks } = await getTasks(
+    {
+      include: {
+        project: { select: { name: true } },
+      },
+      where: {
+        assignees: { some: { contributor: { user: { id: ctx.session.userId as number } } } },
+        status: TaskStatus.NOT_COMPLETED,
+      },
+      orderBy: { id: "desc" },
+    },
+    ctx
+  )
+
+  const upcomingTasks = tasks.filter((task) => {
+    if (task && task.deadline) {
+      return moment(task.deadline).isSameOrAfter(today, "day")
+    }
+    return false
+  })
+
+  // get pastDue
+  const pastDueTasks = tasks.filter((task) => {
+    if (task && task.deadline) {
+      return moment(task.deadline).isBefore(moment(), "minute")
+    }
+    return false
+  })
+
+  return {
+    tasks,
+    upcomingTasks,
+    pastDueTasks,
+  }
+})
